+++ src/packages/grid/demo.taro.tsx
@@ -13,6 +13,7 @@ interface T {
   horizontal: string
   iconStyle: string
   customContent: string
+  event: string
 }
 const GridDemo = () => {
   const [translated] = useTranslate<T>({
@@ -26,6 +27,7 @@ const GridDemo = () => {
       horizontal: '内容横向',
       iconStyle: '图标颜色/大小',
       customContent: '自定义内容',
+      event: '点击子项事件',
     },
     'zh-TW': {
       basic: '基本用法',
@@ -37,6 +39,7 @@ const GridDemo = () => {
       horizontal: '內容橫向',
       iconStyle: '圖標顏色/大小',
       customContent: '自定義內容',
+      event: '點擊子項事件',
     },
     'en-US': {
       basic: 'Basic Usage',
@@ -48,6 +51,7 @@ const GridDemo = () => {
       horizontal: 'Horizontal',
       iconStyle: 'Icon Style',
       customContent: 'Custom Content',
+      event: 'Grid Item Click',
     },
   })
 
@@ -55,6 +59,11 @@ const GridDemo = () => {
     // Toast.text('点击了第几个')
     Taro.showToast({ title: '点击了第几个' })
   }
+
+  const onClick = (item: any, index: number) => {
+    Taro.showToast({ title: `点击了${item.text}，第${index}个` })
+  }
+
   return (
     <>
       <div className="demo">
@@ -157,6 +166,14 @@ const GridDemo = () => {
               icon="https://img12.360buyimg.com/imagetools/jfs/t1/143702/31/16654/116794/5fc6f541Edebf8a57/4138097748889987.png"
             />
           </GridItem>
+        </Grid>
+
+        <h2>{translated.event}</h2>
+        <Grid direction="horizontal" onClick={onClick}>
+          <GridItem icon="dongdong" text={translated.text} />
+          <GridItem icon="dongdong" text={translated.text} />
+          <GridItem icon="dongdong" text={translated.text} />
+          <GridItem icon="dongdong" text={translated.text} />
         </Grid>
       </div>
     </>