+++ src/packages/input/input.taro.tsx
@@ -16,7 +16,7 @@ import { formatNumber } from './util'
 import Icon from '@/packages/icon/index.taro'
 import { useConfig } from '@/packages/configprovider/configprovider.taro'
 
-import { IComponent, ComponentDefaults } from '@/utils/typings'
+import { BasicComponent, ComponentDefaults } from '@/utils/typings'
 
 export type InputAlignType = 'left' | 'center' | 'right' // text-align
 export type InputFormatTrigger = 'onChange' | 'onBlur' // onChange: 在输入时执行格式化 ; onBlur: 在失焦时执行格式化
@@ -30,7 +30,7 @@ export type InputRule = {
 
 export type ConfirmTextType = 'send' | 'search' | 'next' | 'go' | 'done'
 
-export interface InputProps extends IComponent {
+export interface InputProps extends BasicComponent {
   type: InputType
   name: string
   defaultValue: any
@@ -211,9 +211,7 @@ export const Input: FunctionComponent<
   })
   useEffect(() => {
     setClasses(inputClass)
-    if (defaultValue) {
-      SetInputValue(defaultValue)
-    }
+    SetInputValue(defaultValue)
   }, [defaultValue])
 
   useEffect(() => {
@@ -347,7 +345,6 @@ export const Input: FunctionComponent<
     <div
       className={`${classes}  ${className || ''}`}
       style={style}
-      {...rest}
       onClick={(e) => {
         onClick && onClick(e)
         click && click(e)
@@ -439,6 +436,7 @@ export const Input: FunctionComponent<
                 />
               ) : (
                 <input
+                  {...rest}
                   name={name}
                   className="input-text"
                   ref={inputRef}