+++ src/packages/range/range.tsx
@@ -24,12 +24,14 @@ export interface RangeProps {
   hiddenTag: boolean
   min: number | string
   max: number | string
+  minDesc: number | string
+  maxDesc: number | string
+  curValueDesc: number | string
   step: number | string
   modelValue: SliderValue
   button: React.ReactNode
   vertical: boolean
   marks: Record<string, unknown>
-  change?: (value: number) => void
   dragStart?: () => void
   dragEnd?: () => void
   onChange?: (value: number) => void
@@ -69,30 +71,29 @@ export const Range: FunctionComponent<
     button,
     vertical,
     marks,
-    change,
     dragStart,
     dragEnd,
     onChange,
     onDragStart,
     onDragEnd,
+    minDesc,
+    maxDesc,
+    curValueDesc,
   } = { ...defaultProps, ...props }
 
   let { min, max, step } = { ...defaultProps, ...props }
   min = Number(min)
   max = Number(max)
   step = Number(step)
-
   const [buttonIndex, SetButtonIndex] = useState(0)
   const [initValue, SetInitValue] = useState<number | number[] | any>()
-
   const [dragStatus, SetDragStatus] = useState('start' || 'draging' || '')
   const touch = useTouch()
   const root = useRef<HTMLDivElement>(null)
-
   const [marksList, SetMarksList] = useState([])
 
   useEffect(() => {
-    if (modelValue) {
+    if (typeof modelValue === 'number') {
       if (!range && (modelValue < min || modelValue > max)) {
         SetInitValue(0)
         Toast.text(`${modelValue} ${locale.range.rangeText}`)
@@ -270,7 +271,6 @@ export const Range: FunctionComponent<
     }
 
     if ((marks || end) && !isSameValue(value, startValue)) {
-      change && change(value)
       onChange && onChange(value)
     }
   }
@@ -369,7 +369,7 @@ export const Range: FunctionComponent<
 
   return (
     <div className={`${containerName}`}>
-      {!hiddenRange ? <div className="min">{+min}</div> : null}
+      {!hiddenRange ? <div className="min">{minDesc || +min}</div> : null}
       <div
         ref={root}
         style={wrapperStyle()}
@@ -434,7 +434,9 @@ export const Range: FunctionComponent<
                   {button || (
                     <div className="nut-range-button" style={buttonStyle()}>
                       {!hiddenTag ? (
-                        <div className="number">{curValue(index)}</div>
+                        <div className="number">
+                          {curValueDesc || curValue(index)}
+                        </div>
                       ) : null}
                     </div>
                   )}
@@ -469,7 +471,7 @@ export const Range: FunctionComponent<
               {button || (
                 <div className="nut-range-button" style={buttonStyle()}>
                   {!hiddenTag ? (
-                    <div className="number">{curValue()}</div>
+                    <div className="number">{curValueDesc || curValue()}</div>
                   ) : null}
                 </div>
               )}
@@ -477,7 +479,7 @@ export const Range: FunctionComponent<
           )}
         </div>
       </div>
-      {!hiddenRange ? <div className="max">{+max}</div> : null}
+      {!hiddenRange ? <div className="max">{maxDesc || +max}</div> : null}
     </div>
   )
 }