+++ CHANGELOG.md
@@ -1,3 +1,90 @@
+# v1.4.0
+`2022-12-21`
+
+* :sparkles: feat: 新增动态主题定制能力 (#546) @oasis-cloud
+* :sparkles: feat: 新增 input 组件对小程序特性的支持 (#542) @simpleStyle
+* :sparkles: feat: 删除 react-router-dom 依赖，涉及的 to 属性不再支持 (#550)  @oasis-cloud
+* :bug: fix: 修复 searchBar 组件 onSearch 方法。 (#549) @xiaoyatong
+* :bug: fix: 修复 noticbar 垂直滚动延迟问题 (#545) @vickyYe
+
+
+# v1.3.15
+`2022-12-14`
+
+* :sparkles: feat: 新增 animate 动画/动效组件 (#541) @junjun666
+* :sparkles: feat: 新增 ellipsis 组件 (#515) @vickyYe
+* :sparkles: feat: 新增维吾尔语言包 (#526) @MirGhojam
+* :bug: fix: 修复 publish 脚本执行时机错误的问题 (#540) @oasis-cloud
+* 📖 docs: 文档更新 tabbar 的 to 增加属性的说明 (#539)  @oasis-cloud
+
+# v1.3.14
+`2022-12-09`
+
+* :sparkles: feat: image 适配 taro (#507) @vickyYe
+* :sparkles: feat: Grid 组件添加 onClick 事件，统一管理子项点击事件，无需在GridItem组件中分别添加点击事件 (#492) @cuicuiworld
+* :bug: fix: react18 render 提示错误信息 (#524) @oasis-cloud
+* :bug: fix: noticebar 纵向滚动渲染慢问题修复，设置高度小于默认值时动效失效问题修复 (#528) @vickyYe
+* :bug: fix: icon 组件 mask-close 未展示叉号 (#527) @oasis-cloud
+* :bug: fix: #508 修改折叠面板组件修改后不刷新问题 (#521) @zhenyulei
+* :bug: fix: imagepreview 设置 initNo 后  active 计算错误 (#512) @oasis-cloud
+* :bug: fix: 移除 formitem 对 react-router-dom 的依赖 (#513) @oasis-cloud
+* :bug: fix: trendarrow 组件文档增加taro引入方式说明、css文件引入icon的scss文件、修改icon组件的引入方式 (#502) @irisSong
+* 📖 docs: 修正 virtualList 的文档错误 (#510) @Antee
+* 📖 docs: Readme 中英文修改 (#525) @vickyYe
+* 📖 docs: 文档中增加 nutui-react-taro的 import 方式 (#522) @oasis-cloud
+
+
+# v1.3.13
+`2022-11-30`
+* :sparkles: feat: 新增image组件，适配多语言，增加单元测试 (#498) @vickyYe
+* :sparkles: feat: trendarrow趋势箭头组件开发完成 (#491)  @irisSong
+* :bug: fix: 提交skeleton骨架屏白底修改 (#499) @junjun666
+* 📖 docs: pagination demo 增加白色底 (#496) @oasis-cloud
+* 📖 docs: popover demo 增加白色底 (#497) @oasis-cloud
+* :bug: fix: cell 空链接不需跳转 (#495) @oasis-cloud
+* :sparkles: feat: 新增cascader的tabs底部选中颜色 (#493) @simpleStyle
+* :bug: fix: taro demo 启动报 declare 错误 (#490) @oasis-cloud
+* :sparkles: feat: 增加 triangle 图标 (#489) @oasis-cloud
+* :sparkles: chore: interface 统一 @oasis-cloud
+* :sparkles: feat: cascader组件支持activeColor选中态属性 (#487) @simpleStyle
+
+# v1.3.12
+`2022-11-23`
+
+* :bug: fix: children 无数据时不展示cell，应在顶层判断 (#482) @xiaoyatong
+* :sparkles: chore: pulltorefresh 删除取用 console (#480) @oasis-cloud
+* :bug: fix: taro:h5启动时textare组件change时间获取数据报错问题修复 (#479) @vickyYe
+* :sparkles: feat: add form (#470) @xiaoyatong
+* :bug: fix: taro 包不需要依赖 react-router-dom (#475) @oasis-cloud
+* :bug: fix: calendar taro中Popup组件与h5中属性不一致问题 (#468) @Phyzait
+* :bug: fix: Cell组件当children为空数组时不渲染Cell组件问题 (#466) @cuicuiworld
+* :sparkles: feat: range组件新增描述属性 (#471) @simpleStyle
+* :bug: fix: textarea组件字数统计不准确问题修改，兼容emoji (#467) @vickyYe
+
+# v1.3.11
+`2022-11-16`
+
+* :bug: fix: CheckboxGroup和RadioGroup组件动态改变options没有导致子组件更新问题 (#464) @cuicuiworld
+* :sparkles: feat: avatar组件支持iconSize自定义图标大小 (#461) @cuicuiworld
+* :bug: fix: avatar img 标签增加 avatar-img (#462) @oasis-cloud
+* :bug: fix: button 组件中 icon 单独使用不居中  @oasis-cloud
+* :bug: fix: tag组件去除多余标签元素 (#459) @cuicuiworld
+* :bug: fix: calendar 今日标签样式错误 (#460) @Phyzait
+* :sparkles: feat: infiniteloading 组件适配 taro (#458) @junjun666
+* :bug: fix: 修复taro原生img标签无法重新渲染的问题 (#455) @junjun666
+* :sparkles: feat: swiperItem组件支持className和style属性 (#454) @cuicuiworld
+* :bug: fix: 修复TabBar组件demo演示页路由跳转失败问题 (#451) @cuicuiworld
+* :bug: fix: 修复Row组件的文档示例错误 (#453) @simpleStyle
+* 📖 docs: 修复 datepicker 文档中示例代码错误 (#449) @xxxnine
+* :sparkles: feat: 补全button的关于小程序API的typescript支持 (#447) @simpleStyle
+* :bug: fix: input组件defaultvalue值为空时不可清除问题修复 (#452) @vickyYe
+* :sparkles: feat: menuitem 支持 ref (#450) @oasis-cloud
+* :sparkles: feat: badge组件支持className和style属性 (#444) @cuicuiworld
+* :bug: fix: 引入样式文件运行报错(#439) (#441) @hydeia
+* :sparkles: feat: button组件支持原生小程序和H5的button特性 (#433) @simpleStyle
+* :sparkles: feat: row和col组件支持className和style属性 (#440) @cuicuiworld
+* :sparkles: feat: rate组件添加className和style属性 (#438) @cuicuiworld
+
 # v1.3.10
 `2022-11-09`